When no gui event loop is running, propagate callback exceptions.
CallbackRegistry currently defaults to suppressing exceptions that occur
in callbacks because they cause PyQt5 to abort() the program with no
chance of even catching the exception, but such a behavior is annoying
when writing tests involving callbacks (e.g., mplcursors), because in
that case we actually want to see whether an exception occurred, and
typically fail the test in that case.

Instead we can detect whether a GUI event loop is currently running and
propagate the exception if none is running.

Part of the patch is just moving `_get_running_interactive_framework`
from `backends` to `cbook`, so that `cbook` keeps its property of being
importable independent of the rest of mpl.
