+++ doc/api/next_api_changes/behaviour.rst
@@ -17,3 +17,11 @@ Previously, it did not necessarily have such an attribute.  A check for
 ``hasattr(figure.canvas, "manager")`` should now be replaced by
 ``figure.canvas.manager is not None`` (or ``getattr(figure.canvas, "manager", None) is not None``
 for back-compatibility).
+
+`.cbook.CallbackRegistry` now propagates exceptions when no GUI event loop is running
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+`.cbook.CallbackRegistry` now defaults to propagating exceptions thrown by
+callbacks when no interactive GUI event loop is running.  If a GUI event loop
+*is* running, `.cbook.CallbackRegistry` still defaults to just printing a
+traceback, as unhandled exceptions can make the program completely ``abort()``
+in that case.
